pixelRatio: Update tooltips on the General Settings page
The tooltips displayed on the General Settings page have been modified. The previous "Done" prompt has been changed to "Reopen to take effect" to more clearly indicate that users must restart the app after changing settings for the changes to take effect. This change is intended to improve the user experience and make the tooltips more accurate and easier to understand.

Change-Id: Ie4b9292dae221355f5a66f35df9f431dbbb9f0fc