Add a TODO comment to the HomeState class
In the home.dart file, we've added a TODO comment to the HomeState class, noting the need to support removing pieces from the opponent's hand. This change serves as a reminder to the development team that this functionality will need to be implemented in the future to enhance the integrity of the game rules.

This change helps improve code maintainability and ensures that this important functionality isn't overlooked during future development.

Change-Id: Ic1fbe351eeb9eea63de4538fa0e7343838669657