Optimize removal rules to handle special capture modes
When generating removal actions, ensure that general removals are only allowed when there are no special capture modes (such as intervention and guardianship). Updated logic to avoid performing draw removals and all-in-one removals when special capture is available. Additionally, ensure that illegal or failed moves are skipped during the search process, enhancing code robustness.

This change aims to improve the accuracy and consistency of the game engine and ensure that capture mode priority is correctly handled.

Change-Id: I42645436089207da958d37060b8a542681457682