Enhance release script to restrict releases to master or stable branches
This update adds a check in the release script to ensure that releases can only be made from the 'master' or 'stable' branches. It also modifies the push confirmation prompt to reflect the current branch instead of defaulting to 'master'. This improves the safety and clarity of the release process.

Change-Id: I10c5cb9173f2068f799aaa41a57337ff948b655b