Add tag format validation to release script
This update enhances the release script by adding a check to ensure that the latest tag follows the vX.Y.Z format. If the tag is missing or does not conform to this format, an error message is displayed, preventing potential release issues. This improves the reliability of the release process.

Change-Id: I10c5cb9173f2068f799aaa41a57337ff948b655b